Move models to new py file
