+++ modules/python/src/clear.py
@@ -1,2 +1,14 @@
+import json
+from model import DynaSecGroups
+import helper
+from dateutil import parser
+from datetime import datetime, timedelta
+
+
 def handler(event=None, context=None):
+    # dyna_sec_groups = DynaSecGroups()
+    #
+    # now = datetime.now()
+    # parser.parse('2017-12-19T04:59:01.025865+0000')
+    # 2017-12-19T04:59:01.025865+0000
     return event
