+++ outputs.tf
@@ -12,6 +12,13 @@ output "execution_resources" {
 output "lambda_names" {
   value = [
     "${module.lamda_authorize.fn_name}",
-    "${module.lamda_revoke.fn_name}"
+    "${module.lamda_revoke.fn_name}",
+    "${module.lamda_clear.fn_name}"
+  ]
+}
+
+output "events" {
+  value = [
+    "${aws_cloudwatch_event_rule.clear.arn}"
   ]
 }
